# backend/docs/swagger.yaml
definitions:
...
  models.RequestEvent:
    properties:
      approval_date:
        type: string
      approver_id:
        type: integer
      book_id:
        type: string
      reader_id:
        type: integer
      req_id:
        type: integer
      request_date:
        type: string
      request_type:
        type: string
    type: object
  models.User:
    properties:
      contact_number:
        type: string
      email:
        type: string
      id:
        type: integer
      lib_id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        description: Role can be "LibraryOwner", "LibraryAdmin", "Reader"
        type: string
    type: object
info:
  contact: {}
paths:
...
  /requests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RequestEvent'
            type: array
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
  /requests/{id}:
    get:
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RequestEvent'
        "404":
          description: '{"error": "Request not found"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
  /requests/{id}/approve:
    post:
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Approver Data
        in: body
        name: approver
        required: true
        schema:
          $ref: '#/definitions/models.RequestEvent'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Issue request approved", "issue": models.IssueRegistry}'
          schema:
            type: object
        "400":
          description: '{"error": "Invalid request body"}'
          schema:
            type: object
        "404":
          description: '{"error": "Request not found"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
  /requests/{id}/reject:
    post:
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Issue request rejected"}'
          schema:
            type: object
        "404":
          description: '{"error": "Request not found"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
  /requests/pending:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RequestEvent'
            type: array
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "User deleted"}'
          schema:
            type: object
        "404":
          description: '{"error": "User not found"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: '{"error": "User not found"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: '{"error": "error message"}'
          schema:
            type: object
        "404":
          description: '{"error": "User not found"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
      summary: Update a user
      tags:
      - users
  /users/admin:
    post:
      consumes:
      - application/json
      description: Create a new admin user with role "LibraryAdmin"
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: '{"error": "error message"}'
          schema:
            type: object
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new admin
      tags:
      - users
  /users/admins:
    get:
      consumes:
      - application/json
      description: Retrieve all admin users with role "LibraryAdmin"
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: '{"error": "error message"}'
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all admins
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"